requirements.txt
django==5.1.0
gunicorn==23.0.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
google-auth==2.36.0
google-auth-oauthlib==1.2.1
django-cors-headers